Add initial API functions for weather data retrieval
