+++ src/module_1/models.py
@@ -1,0 +1,7 @@
+from typing import Dict, List, Any, Optional, Union, Tuple
+import pandas as pd
+
+
+# Type aliases for better readability
+JSONType = Dict[str, Any]
+DataFrameType = pd.DataFrame
